reuse the module to calculate kl threshold
